UnfoldBIDS._load_results — Function

_load_results(files_df; generate_Xs::Bool = true)

Internal function to load Unfold models into memory. Can also be used to load data after file information was loaded lazily (lazy=true) using load_results()

UnfoldBIDS.list_all_paths — Method

list_all_paths(path)

Internal function to find pathfiles

UnfoldBIDS.load_results — Method

function load_results(bids_root::String;
    derivatives_subfolder::String="Unfold",
    lazy::Bool=false,
    generate_Xs::Bool = true,
    ses::Union{Nothing,AbstractString}=nothing,
    task::Union{Nothing,AbstractString}=nothing,
    run::Union{Nothing,AbstractString}=nothing)

Load Unfold models existing in a derivatives_subfolder in your BIDS root folder.

Keywords

derivativessubfolder (String::"Unfold"): Defines in which subfolder of bidsroot/derivatives to look for Unfold models. lazy (Bool::false): Do not actually load the dataset into memore if true, only return a dataframe with paths generate_Xs (Bool::true): Do not recreate the designmatrix; improves loading time. ses (Union{Nothing,AbstractString}::nothing): Which session to load; loads all if nothing task (Union{Nothing,AbstractString}::nothing): Which task to load; loads all if nothing run (Union{Nothing,AbstractString}::nothing): Which run to load; loads all if nothing

UnfoldBIDS.save_results — Method

save_results(results::DataFrame, bids_root::String; 
    derivatives_subfolder::String="Unfold",
    overwrite::Bool=false)

Function to save unfold models in your BIDS root folder. Automatically creates a derivativessubfolder (default = "Unfold") in the derivatives and subsequentely safes each model in results according to BIDS. Example of path so saved file: bidsroot/derivatives/Unfold/sub-XXX/eeg/sub-XXXses-XXtask-XXXrun-XXunfold.jld2

Keywords

derivatives_subfolder (String::"Unfold"): Creates the named subfolder and saves Unfold models according to BIDS. overwrite (Bool::false): Does not overwrite existing datasets; can be set to true.

UnfoldBIDS.unpack_results — Method

unpack_results(results_df)

Unpack all results into one tidy dataframe/ coeftable.